+++ info/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - LAPS check performed only if machine is domain-joined.
+- Domain membership tested using the Windows API NetWkstaGetInfo rather than the registry.
 
 ## 2024-06-20